requirements.txt

numpy
tensorflow
pillow
scikit-learn
opencv-python
flask
tqdm
